add baseImageProcesor
