+++ mlx_vlm/models/base.py
@@ -1,0 +1,32 @@
+
+from abc import ABC, abstractmethod
+from typing import Dict
+
+from transformers.image_processing_utils import get_size_dict
+from transformers.image_utils import (ChannelDimension, PILImageResampling )
+
+
+class BaseImageProcessor(ABC):
+    def __init__(self,
+                 image_mean=(0.5, 0.5, 0.5),
+                 image_std=(0.5, 0.5, 0.5),
+                 size=(384, 384),
+                 crop_size: Dict[str, int] = None,
+                 resample=PILImageResampling.BICUBIC,
+                 rescale_factor=1 / 255,
+                 data_format=ChannelDimension.FIRST):
+        crop_size = crop_size if crop_size is not None else {"height": 384, "width": 384}
+        crop_size = get_size_dict(crop_size, default_to_square=True, param_name="crop_size")
+
+        self.image_mean = image_mean
+        self.image_std = image_std
+        self.size = size
+        self.resample = resample
+        self.rescale_factor = rescale_factor
+        self.data_format = data_format
+        self.crop_size = crop_size
+
+    @abstractmethod
+    def preprocess(self, images):
+        pass
+
